Store and read energy analysis

+++ Source/AnalyserEnergy.hpp
@@ -22,7 +22,7 @@ public:
     
     ~AnalyserEnergy() {}
     
-    void analyse(juce::AudioBuffer<float>* audio, int& energy);
+    void analyse(juce::AudioBuffer<float>* audio, float& energy);
     
 private:
